+++ tests/functional/CpCest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace simplonprod\googlerecaptchatests\functional;
+namespace juban\googlerecaptchatests\functional;
 
 use Craft;
 use craft\elements\User;
 use FunctionalTester;
-use simplonprod\googlerecaptcha\GoogleRecaptcha;
+use juban\googlerecaptcha\GoogleRecaptcha;
 
 class CpCest
 {