+++ tests/unit/ServicesTest.php
@@ -4,11 +4,11 @@
  *
  * Craft CMS Newsletter plugin
  *
- * @link      https://www.simplonprod.co
- * @copyright Copyright (c) 2021 Simplon.Prod
+ * @link      https://github.com/juban
+ * @copyright Copyright (c) 2022 juban
  */
 
-namespace simplonprod\googlerecaptchatests\unit;
+namespace juban\googlerecaptchatests\unit;
 
 use Craft;
 use craft\helpers\Json;
@@ -16,15 +16,15 @@ use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Psr7\Response;
 use GuzzleHttp\Psr7\Utils;
+use juban\googlerecaptcha\GoogleRecaptcha;
+use juban\googlerecaptcha\services\Recaptcha;
 use PHPUnit\Framework\MockObject\MockObject;
 use Psr\Http\Message\RequestInterface;
-use simplonprod\googlerecaptcha\GoogleRecaptcha;
-use simplonprod\googlerecaptcha\services\Recaptcha;
 use UnitTester;
 use yii\web\ForbiddenHttpException;
 
 /**
- * @author    Simplon.Prod
+ * @author    juban
  * @package   Google reCAPTCHA
  * @since     1.0.0
  */